Fix support for switches with more than 8 ports
Some commands (VLAN, 802.1q, port mirroring) use a bitmap to represent
affected ports.  The length of the data will therefore vary as a 16-port
switch will require two bytes (16 bits).  Until now the code assumed
1 byte per parameter (i.e. up to 8 ports)

Parsing queried data is not too difficult as the code can work out the
length of each parameter based on the length of the received data.

Setting data is trickier as the command has to know how many ports the
switch supports but the code was written to be stateless.

If one of the affected commands is detected (for setting), the number
of ports is first queried and then passed on to all the affected
commands.